+++ public/images/png/icons8-file-96.pngZone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+HostUrl=https://icons8.com/icon/jM97zjbpG515/file
