+++ src/shared/utils/formatMsgramChart.ts
@@ -1,3 +1,4 @@
+import convertToCsv from './convertToCsv';
 import { Historical } from '@customTypes/repository';
 import { format } from 'date-fns';
 import _ from 'lodash';
@@ -73,6 +74,20 @@ const formatMsgramChart = ({ historical, title, isEmpty = false }: Props) => {
     }
   ];
 
+  const handleExportCsv = () => {
+    if (historical) {
+      const csvContent = convertToCsv(historical);
+
+      const blob = new Blob([csvContent], { type: 'text/csv' });
+      const url = window.URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.href = url;
+      a.download = 'dados.csv';
+      a.click();
+      window.URL.revokeObjectURL(url);
+    }
+  };
+
   return {
     title: {
       text: title
@@ -80,6 +95,18 @@ const formatMsgramChart = ({ historical, title, isEmpty = false }: Props) => {
     tooltip: {
       show: !isEmpty,
       trigger: 'axis'
+    },
+    toolbox: {
+      feature: {
+        myCustomTool: {
+          show: true,
+          title: 'Export CSV',
+          icon: 'image:///images/png/iconCsv.png',
+          onclick: () => {
+            handleExportCsv();
+          },
+        }
+      }
     },
     dataZoom,
     grid,
